+++ lib/kb_phylogenomics/kb_phylogenomicsServer.py
@@ -342,6 +342,10 @@ class Application(object):
                              name='kb_phylogenomics.view_tree',
                              types=[dict])
         self.method_authentication['kb_phylogenomics.view_tree'] = 'required'  # noqa
+        self.rpc_service.add(impl_kb_phylogenomics.trim_speciestree_to_genomeset,
+                             name='kb_phylogenomics.trim_speciestree_to_genomeset',
+                             types=[dict])
+        self.method_authentication['kb_phylogenomics.trim_speciestree_to_genomeset'] = 'required'  # noqa
         self.rpc_service.add(impl_kb_phylogenomics.run_DomainAnnotation_Sets,
                              name='kb_phylogenomics.run_DomainAnnotation_Sets',
                              types=[dict])